requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1

pydantic==2.7.4

httpx==0.27.0

openai==1.35.10
anthropic==0.34.0

python-dotenv==1.0.1